Portrait fixes, some warnings removed
git-svn-id: svn://gtkevemon.battleclinic.com/GTKEVEMon/trunk/gtkevemon@6 6e31b030-b45c-3146-9160-cf88ccb029a2

+++ src/serverlist.cc
@@ -66,7 +66,7 @@ ServerList::store_to_config (void)
 
 void
 ServerList::add_server (std::string const& name,
-    std::string const& host, int port)
+    std::string const& host, uint16_t port)
 {
   ServerList::list.push_back(Server(name, host, port));
 }
